Fix missing mesh bug and assign vertex bug
There were two remaining issues visible at the time of the previous
commit, which are both fixed in this commit.

Firstly, meshes were missing after a couple of splitting iterations,
which was due to them being deleted in Terrain::splitLargeMeshes() even
if the splitting was unsuccessful. This was corrected.

Furthermore, the reason the first issue was noticed at all was that
polygon splitting was performed iteratively until a vertex ran out of
poly links, after which splitting failed (since a polygon could not be
added). The cause was a missing 'break;' statement in
Bundle::splitAssignSpikeVertices().

I will probably follow up with a commit for cleaning up excessive
(currently greyed out) test print statements, after which the adjacency
is supposed to return to the main stage.