Check meshes for consistency, suppress text output
In this commit I add a third check for determining the correctness of
the meshes and strips (TopologicalMesh::checkPolyIndices()), and I
greyed out most of the text output that Strata generates in order to
make the important info stand out better.

There are still additional checks that could conceivably help detecting
bad meshes, in particular related to detecting weird topologies that are
not isomorphic to the unit disc.