Fix adjacency bug the *right* way
I spent the last three days trying to fix adjacency, and it took so long
primarily because I went about it the wrong way.

When trying to see why I had issues with Bundles not being seen as
adjacent to Strips, I realized that some Strip vertices
were already referencing their new Bundle. I assumed (oh dear) that this
meant I should also check for membership of the new Bundle and if so,
adjust the remote index anyways and mark the Bundle as adjacent to the
Strip.

What I should have realized was that membership of the new Bundle also
implied having the corresponding remote index, and that all I had to do
is set the flag for being adjacent to 'true' if the vertex already
referenced the Bundle.

Instead, I changed the remote index in both cases, leading to randomly
reassigned remote indices. Not much is lost there, since the position
and polygon structure doesn't change (so the mesh still looked fine) but
the reassignment sometimes resulted in referring to indices in the other
mesh, such that in consecutive splittings we would end up referring to
random vertices of some distant Bundle.

Summarizing, this commit doesn't change much beyond two lines in
Strip::updateAdjacentBundle to ensure properly marking a mesh as
adjacent. And a whole lot of comment lines that I needed to find the
problem, which are going to be gone in the next commit.