Add function TopologicalMesh::checkTopology()
In this commit I wrote a function for verifying a couple of basic
requirements on the topology of the mesh, that should be satisfied by
every Strip and every Bundle.

This mainly includes the possibility to find a trajectory on the mesh by
consecutively finding neighboring vertices, such that exactly the right
number of neighboring polygons (i.e. the number of nonzero entries of
the Vertex::poly array) is found.

This corresponds to finding a set of polygons that is isomorphic to the
unit disc (or half of it), that contains all of a vertex's polygons, and
that doesn't incorporate any polygons not known to the vertex. The
function findPolygon() works such that it uses the vertex's own poly
array, so that this procedure cannot succeed unless all of the vertex's
polygons are neatly positioned.