Increase mesh refinement, add Strip polygon check
In this commit I greatly increased the number of vertices of which the
mesh consists, in order to generate a more realistically sized mesh that
would actually be used in the future by Strata. As far as I can see it
is still generated rapidly, so there is no reason yet to believe that
the procedures developed to cut meshes (and check their validity) are
too computationally expensive.

Also, I added a polygon check to Strip::checkAdjacentMeshes which
verifies that the Strip does not have any polygons for which all three
vertices originate in the same Bundle. Although both the Strip and the
Mesh could be technically valid while this is the case, we should avoid
making the mesh unnecessarily complex. If a polygon has all three
vertices from the same Bundle, then by all means it should belong to
that Bundle, not to a Strip (which is designed to connect distinct
Bundle objects).