Keep correct parentLayer references
Soon I will try to make Strata track several layers on top of each
other. For this, each layer should properly track the layer below it (or
above it), to such an intimate extent that it should be notified of any
mutation on any of the Bundles of the mesh that it depends on. That is
to say, when vertices are added or deleted, dependent Layer objects
should be given the opportunity to adjust their own structure
accordingly.

The Mesh class already had a parentLayer variable, and soon this should
be used in order to establish relationships between different layers.
The plan is to use a single 'master' layer deep below the ground for
generating the primary deformations (i.e. formation of mountains and
valleys), and have all other layers follow this deformation. The layers
on top of the master layer should retain their thickness, but they can
be eroded away, and new layers can be deposited on top of older layers.

I have not yet decided what is the best way to implement this dependency
as this concerns massive numbers of vertices being mapped (in a way) to
other vertices. Using std::map-like structures is probably too slow, but
a static non-indexed map is hard to keep up-to-date with mutations in
the underlying mesh.