Create MasterLayer class
This commit specializes the Layer class into the MasterLayer, a special
class that will be the foundation of our terrain. This layer can be
initialized with a flat layer and then be built upon. No other layer
should have the possibility to be a completely independent layer - all
should somehow be supported by the layers below them.