Correctly track Bundles in Layer class
There was a list of Bundles in the Layer class that was supposed to
contain pointers to all Bundles of the Layer (for the obvious purpose of
doing operations on all vertices of a single layer). The list was hardly
populated, though - and that should now have been corrected.

I have spent a lot of time thinking about how to deal with the
interactions between layers. The end result is that I do not wish to put
anything specific in the Mesh classes - anywhere in the /src/mesh/
directory - to accommodate this. While it seems practical to help
interactions (such as constantness of volume) by adding a notion of
thickness, or connections between vertices of one layer and vertices of
the layer under it, to do so is against the intention of the Mesh
classes to only facilitate (1) rendering of the Terrain, and (2)
managing the topology of the intra-layer boundaries.

That is to say, the Mesh class should make sure that Terrain objects can
be visualized and that there exists a well-defined topological
relationship between the vertices that define a layer boundary. To also
let the Mesh class govern the way in which layers are related to each
other would be a big step, and one that is better left to a different
part of Strata that would focus on the physics of the terrain, its
material properties, its behavior under compression, extension and
shear, etcetera.

There is no reason at this time that strictly requires any relationship
between layers - although I do admit that a fast algorithm is necessary
to see the 'distance' between a vertex from one layer to the layer under
it, since such lookups are going to be required frequently. Therefore I
am going to continue on producing a few more layers, after which I
intend to move on to the parts of Strata which actually modify the
terrain.

Initially that should be about compression and folding, then about
erosion, and finally about sedimentation. The order in which these
things are to be added is dictated by logic - there is no erosion
without height differences, and there is no sedimentation without
erosion. In addition to this, folding is 'easy' because it does not
involve modifying a mesh, and erosion is 'relatively easy' since it
doesn't involve adding new layers or selectively adding to them.