Allow duplication of layers
Today's commit is a bit premature and pushed through primarily because
time ran out and I need to sleep. That said, there is something here
because everything compiles and runs in spite of everything I've tried
to do.

For creating the multi-layer terrain that serves as the initial,
pre-modification terrain I need several layers on top of each other. The
most convenient and logical approach is to learn to copy an existing
Layer, something that turned out to be a bit more technical than it
should have been.

I need to update all the cross-references between Bundles and Strips
after the copying is done, and that's what the current commit is mostly
about. There are still some flaws evidenced by a bit of a hiccup at
startup and a lot of warning spam at shutdown. Also, I need to look at
the code with a fresher mind tomorrow.

When these remaining issues have been addressed, focus should shift
towards terrain modification as mentioned in recent commits. Such should
happen in different sections of Strata that should be well-separated
from the Mesh, although I thought a bit more about this and decided that
I probably do need to keep references from terrain manipulation code to
Mesh vertices to keep layers from sliding over each other (i.e. I need
to pin them down and try to avoid movement, which is forbidden by
intra-layer friction).

The latter should be solved by taking a couple of reference points and
keeping these in line, or by tracking the speed of movement of a layer
and using that as comparison. One way or another, somehow, there needs
to be a subtle logic in making the layers interact properly, but that
solution will take considerable time, effort and creativity, and I hope
I will be able to keep it clean and outside of the Mesh as much as
reasonably possible.