Render and translate duplicated layers
In this commit I add functions for rendering layers that are duplicated,
and I added a way to translate these layers.

The rendering is taken care of mostly by Terrain::duplicateLayer(),
which initializes the mesh objects of the new layer using the Drawable's
new class functions resetTexture() and getTexture().

Translation is done by the Layer on its Bundle objects, and subsequently
for the Strips by Terrain::duplicateLayer() calling the new function
Strip::recalculateVertexPositions(), which re-sets the Strip positions
by re-getting them from the Bundle.

Strata runs but the references of some StripVertex objects to their
remote vertex are clearly off. This is most likely a previously existing
but so far unnoticed bug in the splitting, which will have to be dealt
with in a future commit.

Otherwise, layer duplication itself seems to work adequately and this
should be the last commit to add functionality for this duplication. In
principle one should now be able to develop a flat base terrain with
multiple layers, upon which manipulation would be possible.

The focus should therefore shift towards accommodating such
manipulation, by developing other parts of Strata to work on the cloud
of vertices that the Terrain should be to the outsider. For the Mesh
classes, functions are still required for relaxing the mesh, e.g. by
joining vertices, splitting polygons, et cetera, such that very
elongated polygons are dealt with properly and such that polygons do not
become too large or too small.