Use correct indices when splitting Strips
That was rather subtle.

So in the previous commit there was a visible bug with some Strips
borrowing the wrong vertices from Bundles. It turns out that this was
due to a bug in Mesh::splitAssignPolygonsToConstituentMeshes(). There,
when adding polygons that are not entirely contained within either of
the two meshes that result from the splitting, the Mesh uses a special
function for creating new vertices in the Strip using vertices from the
post-split meshes.

So long as we were splitting Bundles this was fine, but when we were
using vertices from Strips it took the index in the new Strip object and
(via Mesh::getVertexOwner()) applied that index to the Bundle that owned
the vertex. We thus referred to the Bundle's vertex by using an index
intended for the Strip - a rather serious mismatch.

What needed to be done is to add the same logic as getVertexOwner() in
order to get the Vertex's remote index in the owning Bundle. This commit
adds functions for this, and it adjusts the Mesh's split procedure for
mixed polygons such that it consistently uses the owning bundle and
remote index in that bundle. This is now possible by virtue of the
polymorphism of Mesh::getVertexOwner() and Mesh::getRemoteVertexIndex().