Develop a new way to calculate terrain height
The TerrainManager still contained the nowadays ancient getHeight
function that was based off interpolation on a world-scale 2D texture.
Since Strata has a very different structure, and since additional logic
is necessary anyways, I developed a function for determining the height
of Strata's terrain at a certain point.

The functions are considerably more complex than interpolation on a
texture, but I have faith in the possibility for height lookups to
become fast enough not to be a computational issue.

The procedure I developed today (and which is not yet needed for
anything, and therefore is not functional) consists of looking whether
Bundles and Strips of the Terrain are close enough to potentially be
below a given point in space, and then a search through the nearby
Bundles and Strips to find a polygon that is directly beneath it. When
such a polygon is found, the height is the intersection point between
the polygon plane and the vertical axis through the point of interest.

The most basic geometric functions, primarily the one for finding the
intersection point between a plane and a line, went to
/src/mesh/vecmath.hpp, and the topological mesh performs the check on
all polygons it contains. Given that there is no difference between
Bundles and Strips, nothing is changed specifically for these classes.

I also changed the name of the height functions of the TerrainManager to
getVerticalHeight() and similar, because there is obviously reason to
also be interested in collision with vertical terrain features. A simple
getHeight() is therefore ambiguous. Furthermore, one will have to
specify a vec3 rather than a vec2 because Strata explicitly wishes to
allow the existence of warped terrains with caves, tunnels and arches.