Add simple text box to Strata
In this commit I add the font rendering of tiny-game-engine to Strata's
UI manager. It is very simple because I just wanted it to work - for
now. I will build upon it later.

I ran into a bit of trouble because the text wouldn't show up when the
sky was active, and the terrain wouldn't show up when the sky was not
active. It turns out I was adding the text to the Renderer before the
sky - and the Renderer cannot know which of the two should be in front.

That is now solved (by switching the order), but adjustments to
tiny-game-engine might be in order. For now, everything works though,
and I will try to standardize the text rendering a bit more and modify
it into more tidy and structured code.