Use Lua to initialize UI variables
Since the UI testcase was working properly, I proceeded to move the
initialization of UI variables (e.g. which font to use, what the default
font size is, how big a font texture to create) to Lua. Since I
eventually want to set up the entire UI from Lua, this is a logical
move.

Any subsequent UI elements to be added to Strata will be added through
as-generic-as-possible Lua-callable functions, where the exact shape and
composition of the UI is controlled through Lua scripts.