Used Lua to create a Monitor window
It was clearly a while since I last used Lua and Selene, since I kept
getting stuck at trivial issues. But in the end it did work, and now all
the parameters for creating the Monitor window are brought to Lua
scripts. One can now use Lua to set the title (if any is desired), and
toggle whether or not fps are displayed. As the monitor grows, one
should be able to flag what elements should be displayed and what
elements should be hidden.

Today's approach did highlight a few weaknesses in the current setup
that may need some attention in the near future. Firstly, the interface
(e.g. RenderInterface) should be accessible widely, and it is not good
style to include things that look like '#include
"../core/interface/render.hpp"' in all kind of files. I am considering
to move the interface directory up one level, and then make it
accessible to all. For instance, if one wants to display stats about the
Terrain, one needs the UI to render some things that it needs to pull
from there. A messaging system of some kind might be in order.

Secondly, function signatures for C++ object constructors (from Lua) can
get super-long. For instance, if one wants to load a character from Lua
with 60 attributes one might already need a 60-argument function. This
is inefficient. There should be some attempt to encapsulate information
in a bundled way, such that Lua can pass large amounts of information in
just a few objects. Said otherwise, Lua should return struct-like
objects or long decodable text strings, rather than heaps of arguments.

Both of these weaknesses need a delicate solution, and neither appears
to be easy at this stage, so it is quite possible that it takes a while
to address them.