Increase terrain refinement, fix texture resolution
This commit confirms that Chathran Strata scales properly, and that a
terrain consisting of over 2 million vertices (22 layers of 104275
vertices and 207255 polygons each, where each layer measures 1024x1024
and is made of triangles with an edge size of about 3.0) can be rendered
without any observed slowdown.

This commit also fixes a bug due to using scaleTexture (the conversion
factor used to scale from world coordinates to (0,1) texture
coordinates) for the texture resolution (i.e. the number of pixels that
determines how detailed the texture is). No variable existed for the
latter parameter, and it is now added as DrawableMesh::sizeTexture.

Memory consumption is still a bit on the high side (with nearly 1 GB
memory required for a terrain that needs 260mb for its vertices and
polygons), but that is because every mesh still has its own texture.
This isn't necessary - they all use the same texture, so some sharing
could take place. I will try to improve this in future commits.