Reuse texture for all Meshes of a Layer
In this commit I give a single texture object to each Layer, which is
then used for each of the Layer's meshes.

This also implies that the DrawableMesh itself no longer manages its own
mesh, and it does not create or destroy it.

Meanwhile I observe that the memory usage of Strata remains at 1 GB for
260 MB of terrain data, so there has to be some other usage of memory
that might be unnecessary. It could be somewhere in the StaticMesh or
so, but more investigation is in order.