Calculate allocated, not used, memory for Terrain
In this commit I also added using 'capacity' rather than 'size' to
determine how much space is allocated for the Terrain's vertex and
polygon arrays. The difference is significant, with 420 mb used but 560
mb allocated.

Meanwhile, I did the same for the StaticMesh but there the allocated and
used memory is equal (likely because hostData is reserved to be exactly
the required size).

In the end this leaves 560mb of 950mb memory usage 'found'. The
remaining 400mb must be somewhere, it could be the textures, but either
way it requires additional investigation. Separately, using
std::vector::reserve() where possible seems to be worthwhile as it could
apparently reduce memory usage of the vertex/poly arrays by something
like 30% (or 15% of the total currently used by Strata).