Use base class collectArgs in Lua
In Chathran Lua scripts, a collectArgs function is used to assemble the
arguments used to construct objects in C++ via registered C++ functions.

In the case of the Monitor, the Lua object for the Monitor is derived
from a general Monitor object. With the attribute-value specialization
introduced in the previous commit, it is now possible to use only the
generic base class collectArgs() to construct derived objects.

This is of significant importance generally, as it means that base
objects can all be initialized equally with the same Lua functionality
on a different C++ registered function that is specific to the derived
class. This avoids the need for copying the entire collectArgs function
for each derived class, which keeps the Lua scripts much simpler.