Add thickness as Vertex attribute
The matter of tracking a layer's thickness has haunted Strata for
considerable time. In principle, one wants the Mesh classes to define
only the Mesh, and not have any Terrain logic in it. That helps in
keeping the Mesh fast and efficient. On the other hand, the vertices of
the Mesh will be moved around by terrain generation, and if the vertices
are managed entirely by the Mesh then there can be no vertex-specific
management by the terrain generation.

Such specificity is good, in a certain sense, because it forces us to
make terrain generation generic and universally applicable. The terrain
generator knows the nature of each layer and should therefore be able to
tell about each point, locally, where it goes in response to large-scale
forces at work.

In practice, the terrain generator will have to carry out a loose vertex
mapping, by which I mean the following: it maps each vertex and performs
manipulations on the terrain for a while, after which it deletes its
mapping, lets the Mesh-Terrain do the cleaning-up, and then re-creates a
new mapping from scratch (since the Mesh-Terrain will have changed the
structure of all its meshes).

That re-mapping could be done by letting the Mesh-Terrain send a notice
to the Terrain every time it modifies a Mesh, upon which the terrain
generator would have to replicate the move. That is hardly elegant, and
it smells of doing the same operation twice. The terrain generator, in
principle, is intended to only move around vertices, and it should try
to avoid topological questions as much as reasonably possible - the
topology is the Mesh-Terrain's responsibility. By all means the Terrain
should be a cloud of vertices to the terrain generator. The movement of
vertices in the cloud is then governed by the following:
- the movement of nearby vertices;
- the material properties of the layer that the vertex is part of;
- whether or not the vertex is at the surface layer (for
  erosion/deposition);
- the characteristics of its direct neighbours (e.g. for river flow).
All of these questions can be answered properly by the Mesh-Terrain, so
we are almost there. The only issue I see is that in this setup,
intra-layer composition is ill-described: it is not possible for the
terrain generator to see the difference between extensional thinning
(where a layer gets thinner because it is stretched) and erosion (where
a layer gets thinner because its upper parts are removed). If this
difference is not tracked somewhere in Strata, it would become
impossible to determine what the intra-layer structure is, which would
complicate realistic rendering of layers.

This issue can be fixed by tracking the thickness of the layer at each
vertex. For this, an additional 4 bytes per vertex are required, and in
order to keep the total size to 64 bytes the maximal number of edges
next to each polygon is reduced to 10. Given that perfect triangular
meshes have 6 edges for every interior vertex, 10 should be enough,
albeit only barely.

On a related note, one wonders if by any chance the xPoly could be made
a short unsigned int (ideally 2 bytes) rather than 4. This would limit
the maximum number of polygons to 65535, which is not quite high, and
presently Strata already reaches that limit (with over 200k polygons in
each layer).

In spite of the long commit message, the change is rather simple, and
the only place outside of the Vertex class where anything changed is in
the Mesh::splitEdge() function where a new vertex has to be assigned a
thickness equal to the average of the thicknesses of the previous edge
endpoints.