Add UI Interface
In this commit an interface is added for the UI, by means of which the
UI can request information from other parts of Strata for the sake of
displaying on-screen messages, windows et cetera.

Every class that wants to expose its information to the UI can do so by
deriving from intf::UISource and implementing its pure virtual function
intf::UISource::getUIInfo(). The UISource object must be created with a
unique key, which is then used by the UI to find it.

There are some refinements to be done (for instance, ensuring that
UISource objects are created with valid keys), but the principle of how
the UI interface is supposed to be used is proven by this commit's
ability to display the memory usage of the Terrain on-screen.

On a final note, I am afraid the previous commit might not have worked
properly because I did not update CMakeLists.txt to search the correct
directories for source files. However, this commit has a valid
CMakeLists.txt again.