Cut memory usage by conservative reservation
It was observed earlier that the allocated memory of the Terrain is far
larger than the actual used memory. That seems to be because std::vector
doubles its capacity every time it needs to expand beyond its previous
capacity.

I think a doubling is excessive, and in any case the Terrain is such
that, after construction, only a small number of additional vertices and
polygons gets added (relative to the total number). I therefore added
code to reserve 105% of required memory every time the capacity is about
to be exceeded. This does not seem to slow anything down, but it does
bring down the memory allocated (for the Terrain only) from 557mb to
436mb. That is quite a cheap way to reduce memory usage drastically.