Use Lua to generate initial Terrain
I moved the parameters for generating the Terrain, as well as its
layers, to Lua scripts. With this, one of the last hardcoded chunks of
Strata has become a bit more externally controllable.

The size, number of divisions, maximal sub-mesh size, and the thickness
of the layers are now customizable. Eventually, one should be able to
add layers on-the-run from the console, while running Strata.

In the end, the aim is not to add layers like that, but rather to be
completely in control of the Terrain generated by Strata, and then it is
not acceptable for any hardcoded parameters to affect the
characteristics of the Terrain.