Small adjustments to Strata structure
To celebrate Strata's 100th commit, here are a few meaningless
adjustments on the size and colouring of Strata's default terrain. This
mostly concerns changing the Lua script parameters, and changing a few
of the remaining hardcoded values (the camera starting position and the
way the camera rotates and moves). In other words, unimportant
adjustments.

Meanwhile, the justification for this commit is that whatever happens
next is probably going to be something new and big, which I do not want
to mix with the changes committed now.

Two projects are being considered as the next big thing:
1. The addition of Stitches, as the surfaces that connect the edge of a
   Layer to the Layer under it;
2. The implementation of compressional movement along a given tectonic
   movement vector field, and employment of conservation of volume in
   order to generate uplift.

At the time of writing I am considering (1) first, followed by changing
the initial terrain to small continental crust plates surrounded by
ocean floor plate. This is more interesting than doing (2) first because
without (1) there can be no faulting or subduction. There would only be
thickening of the crust, without the possibility for layers to be
detached or fractured, and that isn't very exciting.

In fact, for a realistic situation one would have to go a bit further
and start with a single continent, split it via a rift, and then
re-collide it in order to cause the correct stratigraphy at the edges of
the continental shelf. However, this is probably more of a refinement
and not exactly necessary for the first implementation of orogeny.