Bring back colouring of Strips
In the past Strips were coloured yellow and Bundles were red. This was
eventually abandoned when the Layer came to host the Texture used.
However, for visual feedback on whether Mesh operations are doing the
intended thing the distinction is quite useful.

Since I am going to add Stitches (option (1) from the previous commit) I
am going to have to add a variety of new operations on Mesh objects. In
particular, I am going to delete vertices and move polygons between
Strips and Bundles. Since that isn't likely to work at first try, I
defined three textures (rather than 1) in the Layer. Two of these are
used now - for texturing Bundles and Strips - and the third, the
stitchTexture, will be used for special Strips.

Initially I was going to make a Stitch class for Strips that connect a
Layer to the one under it. However, I do not think there is much of a
reason to make a second class for what is basically very similar to the
Strip in that it connects various borrowed vertices from various Bundles
into a strip-like mesh.