Add StripPolygon, define Stitch interface functions
This commit adds nothing in functionality, but a lot in code.

Firstly I defined a StripPolygon as a loose polygon (i.e not part of any
mesh). The plan is to create Stitches by creating all polygons that are
required to give every edge vertex of the upper Layer a disc-isomorphic
neighborhood (i.e. they need to be in the interior of the Layer, not on
the edge). Said otherwise: every vertex should be surrounded completely
by polygons, and vertices for which this is not the case will have
StitchPolygons created to make it so.

Secondly I moved duplicateLayer() from mesh/terrain.hpp to
mesh/terrain.cpp because it was excessively big. Git does not consider
this a move but a addition-deletion, which likely clutters up this
commit. Pity.

Thirdly I added a centralPoint and a maxDistanceFromCenter to the
TopologicalMesh to speed up searching. Searching for nearby vertices is
rather inefficient (it basically needs to run through all existing
vertices), and by only considering nearby Bundles the search is far more
efficient. I could in principle add a 3D search tree, but that would
increase Strata's already huge memory usage further, so I do not want to
do so. Instead I do two-step searching by skipping Bundles whose
vertices are all too far away to be interesting.

And finally, I added a few functions and variables (like
Terrain::addNewStitch(), Terrain::getUnderlyingVertex() and
TopologicalMesh::isStitchMesh()) that are not used anywhere at this
time, and that I will be using for creating Stitches.