Add functions for identifying Layer edges
The next step in stitching Layers is now written, although it remains
unused since it is still not enough.

The main additions here are the following:
- Bundle::isAtLayerEdge(xVert) tests whether a Vertex is at the Layer's
  edge;
- Strip::findRemoteVertexPolyNeighbor() tries to find a neighbor to a
  given pair of vertices;
- Terrain::stitchLayer() should do the eventual stitching (incomplete).

I am still working on finding the very first vertex-at-layer-edge.
However, when that is found, and the relevant functions are there, the
rest should be 'easy', namely that consists of iteratively finding the
next vertex-at-layer-edge. Then, a nearby vertex should be found on an
underlying Layer (using the to-be-written function
Terrain::getUnderlyingVertex()) and from that point onwards we have to
follow the surface of underlying Layers, and connect the two series of
vertices into a Stitch Strip.

A lot of work remains to be done, therefore, and I guess it's not going
to be done tomorrow either. This commit message should serve as the
blueprint for what the next commits need to deliver.