Enabled that part of stitchLayer that works
In the previous commit it only didn't work because no StartBundle was
found. I now re-enabled the 'safe' parts, but other than that I got
nowhere.

At this time it doesn't look like any of the new layer's vertices are
considered 'at the Layer's edge', so the functions there aren't working
well yet.