Find vertices at Layer's edge
This commit contains the changes necessary to successfully detect
whether vertices are at the edge of the Layer - which is required in
order to perform stitching.

The main subtlety is in looking around in all Strips adjacent to a
Bundle to determine whether there exist polygons for a given Vertex such
that the polygons form a plane for which the Vertex is in the interior
(rather than on the edge).

For this purpose, Bundle::isAtLayerEdge() iterates through Strip
polygons to see if the Vertex is in the interior. Inside the Strip, the
work is done by Strip::findRemoteVertexPolyNeighbor which looks for the
remote neighbor to a remote pair of vertices. This requires the Strip to
first find the local vertices for the given remote vertex indices. After
that, a simple call to the previously existing
TopologicalMesh::findPolyNeighborFromVertexPair() suffices to determine
whether the desired neighbor exists. The starting point is the
counterclockwise along-the-Bundle-edge for the target Vertex.

Since there are multiple adjacent Strips to most Bundles,
isAtLayerEdge() simply looks in all of them and decides that a Vertex is
at the Layer edge if none of the Strips contains the necessary polygon.
On the other hand, if a polygon is found for which the neighbor vertex
equals the clockwise along-the-Bundle-edge vertex, the path through the
adjacent Strips is complete and the Vertex is not at the edge.

On a final note, the previously stated Bundle requirements (no Vertex is
a bottleneck vertex, said otherwise, every Vertex has only 2 edges that
are at the edge of the Bundle; there are no holes in any Bundle; et
cetera) are definitely assumed to apply. If the Bundle requirements on
mesh sanity are violated, there is no guarantee that there will be any
sensical output to the at-Layer's-edge check, and it could probably
give wrong results in both ways.