Extend StripVertex for Stitches
How to deal with Stitches precisely has been vague up to now, and
intentionally so. The main issue that needed a resolution was how to let
the Stitch interact with the Layers beneath it.

Todays commit decides this question in favor of the
complex-but-clean-and-tidy option. Basically, one could do any of the
following things to connect a Stitch to an underlying Layer:
1. Choose an appropriate vertex and connect to it. Switch vertices when
applicable, but do not alter the underlying Layer.
2. Choose an appropriate vertex and move it as the Stitch needs, to keep
the Stitch in the right place.
3. Connect the Stitch to positions along edges of the underlying Layers
instead of to Vertices.
Each option has its disadvantages. For (1), it is not possible to define
e.g. a 'flat' Layer properly. This would make it hard to do things like
erosion or sedimentation at the edges of a Layer. For (2), everything
starts depending on each other and in any case one needs to modify the
underlying Layer to suit the positioning of the overlying one. Keeping
meshes clean and tidy is going to be hard enough anyways, and having to
adhere to vertices remaining at given positions isn't helpful. Moreover,
the idea is to let terrain generation take place through moving around
vertices, which is hard if they are getting moved around by the Mesh
classes as well.
That leaves (3), where the StripVertex needs to be extended to become a
synthetic Vertex somewhere in between two remove Vertices (for Stitch
Strips only). It may become a little harder to handle the Stitch like
that, and some functions become a bit more complicated, but the clean
handling of Stitches that results from it seems to be worth the effort.