Extend Strip::checkAdjacentMeshes() for Stitches
Stitch vertices have a second Bundle, and that one also needs to satisfy
all the requirements of an adjacent Bundle. Additionally, it will now be
possible for Bundles to be adjacent only through the secondary remote
vertex of a Stitch's StripVertex.

In order to accommodate Stitch Strip objects, the previous commits added
some functionality through additional fields of the StripVertex. This
commit completes the adjustments, and all of the Strip class should now
allow the creation of Stitches.