Fix missing Layer bug
It became apparent to me yesterday that the top Layer was not there.
Well it kind of was, since I could see the memory usage and I could see
everything (e.g. Renderables) getting added on inspection, but it wasn't
visible.

It turns out that this was because the moving of vertices and the
setting of textures was in the wrong order. Initially I thought there
was a Layer really *missing*, since I expected that if a Layer was drawn
together with the masterLayer at a different color, it would be visible
through flickering due to equal positioning. Somehow (maybe because they
are truly identical) that did not happen, and I looked for a long time
for a Layer that was not drawn at all, and then (after I discovered the
colour seemed to be off as well) I tried to find out where Layers were
given the wrong color.

In the end there are two contributors to the misery. Firstly,
Terrain::duplicateLayer() should not create meshes until all the vertex
positions have been correctly set. Secondly, if the vertices somehow do
move after mesh creation, then one should use the now added
Drawable::resetMesh() and not Drawable::resetTexture() since the latter
does not actually change anything about the mesh, it only resets the
texture (as its name would suggest).

Besides these corrections, I made the references to Bundles and Strips
to be duplicated 'const' to prevent any accidental manipulation of the
base Layer.