Add isTransverseStitch flag to Strip
Stitching is a nice idea, but it solves two distinct problems: layers
ending because they are deposited only locally (i.e. their thickness
reaches 0 at some point), or layers ending because they have been cut
away beyond some point (i.e. their interior is visible due to a cutoff).

The texturing is different in both cases, and in all likelihood the
terrain generation should be different as well (for instance,
sedimentation result in the first kind of layers, which would gradually
thicken, whereas erosion results in the second and typically leads to
shortening of the Layer by a regressing boundary).

Therefore I added a parameter to the Strip to denote this. Note that it
cannot be in the Layer, since a Layer can have both types of Stitch
Strips.