Continue work on Stitches
This commit mostly contains a bit more additions to the Terrain for
Stitching. I detailed the outline for the Stitching procedure in the
comments for Terrain::stitchLayerTransverse() which is intended for
stitching layers transversely onto the underlying terrain.

Separately, I realised that splitting Stitches will not work right away,
so I inserted a check that warns that this is not yet possible. It
should become possible in the future, but I need to think about what is
the best way to do it.